ENH,BUG: Replace `int` with `size_t` in cub function calls
This replaces the `int` size related parameters with `size_t` parameters
in CUB to remove call limitation and fix existing missing check.

Commit created in part with cursor.

+++ cupy/cuda/cupy_cub.cu
@@ -249,11 +249,11 @@ using _multiply = cuda::std::multiplies<>;
 struct _arange
 {
     private:
-        int step_size;
+        size_t step_size;
 
     public:
-    __host__ __device__ __forceinline__ _arange(int i): step_size(i) {}
-    __host__ __device__ __forceinline__ int operator()(const int &in) const {
+    __host__ __device__ __forceinline__ _arange(size_t i): step_size(i) {}
+    __host__ __device__ __forceinline__ size_t operator()(const size_t &in) const {
         return step_size * in;
     }
 };
@@ -743,7 +743,7 @@ struct select_min<__half> {
 struct _cub_reduce_sum {
     template <typename T>
     void operator()(void* workspace, size_t& workspace_size, void* x, void* y,
-        int num_items, cudaStream_t s)
+        size_t num_items, cudaStream_t s)
     {
         DeviceReduce::Sum(workspace, workspace_size, static_cast<T*>(x),
             static_cast<T*>(y), num_items, s);
@@ -753,7 +753,7 @@ struct _cub_reduce_sum {
 struct _cub_segmented_reduce_sum {
     template <typename T>
     void operator()(void* workspace, size_t& workspace_size, void* x, void* y,
-        int num_segments, seg_offset_itr offset_start, cudaStream_t s)
+        size_t num_segments, seg_offset_itr offset_start, cudaStream_t s)
     {
         DeviceSegmentedReduce::Sum(workspace, workspace_size,
             static_cast<T*>(x), static_cast<T*>(y), num_segments,
@@ -767,7 +767,7 @@ struct _cub_segmented_reduce_sum {
 struct _cub_reduce_prod {
     template <typename T>
     void operator()(void* workspace, size_t& workspace_size, void* x, void* y,
-        int num_items, cudaStream_t s)
+        size_t num_items, cudaStream_t s)
     {
         _multiply product_op;
         // the init value is cast from 1.0f because on host __half can only be
@@ -780,7 +780,7 @@ struct _cub_reduce_prod {
 struct _cub_segmented_reduce_prod {
     template <typename T>
     void operator()(void* workspace, size_t& workspace_size, void* x, void* y,
-        int num_segments, seg_offset_itr offset_start, cudaStream_t s)
+        size_t num_segments, seg_offset_itr offset_start, cudaStream_t s)
     {
         _multiply product_op;
         // the init value is cast from 1.0f because on host __half can only be
@@ -798,7 +798,7 @@ struct _cub_segmented_reduce_prod {
 struct _cub_reduce_min {
     template <typename T>
     void operator()(void* workspace, size_t& workspace_size, void* x, void* y,
-        int num_items, cudaStream_t s)
+        size_t num_items, cudaStream_t s)
     {
         if constexpr (std::numeric_limits<T>::has_infinity)
         {
@@ -817,7 +817,7 @@ struct _cub_reduce_min {
 struct _cub_segmented_reduce_min {
     template <typename T>
     void operator()(void* workspace, size_t& workspace_size, void* x, void* y,
-        int num_segments, seg_offset_itr offset_start, cudaStream_t s)
+        size_t num_segments, seg_offset_itr offset_start, cudaStream_t s)
     {
         if constexpr (std::numeric_limits<T>::has_infinity)
         {
@@ -841,7 +841,7 @@ struct _cub_segmented_reduce_min {
 struct _cub_reduce_max {
     template <typename T>
     void operator()(void* workspace, size_t& workspace_size, void* x, void* y,
-        int num_items, cudaStream_t s)
+        size_t num_items, cudaStream_t s)
     {
         if constexpr (std::numeric_limits<T>::has_infinity)
         {
@@ -871,7 +871,7 @@ struct _cub_reduce_max {
 struct _cub_segmented_reduce_max {
     template <typename T>
     void operator()(void* workspace, size_t& workspace_size, void* x, void* y,
-        int num_segments, seg_offset_itr offset_start, cudaStream_t s)
+        size_t num_segments, seg_offset_itr offset_start, cudaStream_t s)
     {
         if constexpr (std::numeric_limits<T>::has_infinity)
         {
@@ -906,7 +906,7 @@ struct _cub_segmented_reduce_max {
 struct _cub_reduce_argmin {
     template <typename T>
     void operator()(void* workspace, size_t& workspace_size, void* x, void* y,
-        int num_items, cudaStream_t s)
+        size_t num_items, cudaStream_t s)
     {
         DeviceReduce::ArgMin(workspace, workspace_size, static_cast<T*>(x),
             static_cast<KeyValuePair<int, T>*>(y), num_items, s);
@@ -921,7 +921,7 @@ struct _cub_reduce_argmin {
 struct _cub_reduce_argmax {
     template <typename T>
     void operator()(void* workspace, size_t& workspace_size, void* x, void* y,
-        int num_items, cudaStream_t s)
+        size_t num_items, cudaStream_t s)
     {
         DeviceReduce::ArgMax(workspace, workspace_size, static_cast<T*>(x),
             static_cast<KeyValuePair<int, T>*>(y), num_items, s);
@@ -936,7 +936,7 @@ struct _cub_reduce_argmax {
 struct _cub_inclusive_sum {
     template <typename T>
     void operator()(void* workspace, size_t& workspace_size, void* input, void* output,
-        int num_items, cudaStream_t s)
+        size_t num_items, cudaStream_t s)
     {
         DeviceScan::InclusiveSum(workspace, workspace_size, static_cast<T*>(input),
             static_cast<T*>(output), num_items, s);
@@ -949,7 +949,7 @@ struct _cub_inclusive_sum {
 struct _cub_inclusive_product {
     template <typename T>
     void operator()(void* workspace, size_t& workspace_size, void* input, void* output,
-        int num_items, cudaStream_t s)
+        size_t num_items, cudaStream_t s)
     {
         _multiply product_op;
         DeviceScan::InclusiveScan(workspace, workspace_size, static_cast<T*>(input),
@@ -964,7 +964,7 @@ struct _cub_histogram_range {
     template <typename sampleT,
               typename binT = typename std::conditional<std::is_integral<sampleT>::value, double, sampleT>::type>
     void operator()(void* workspace, size_t& workspace_size, void* input, void* output,
-        int n_bins, void* bins, size_t n_samples, cudaStream_t s) const
+        size_t n_bins, void* bins, size_t n_samples, cudaStream_t s) const
     {
         // Ugly hack to avoid specializing complex types, which cub::DeviceHistogram does not support.
         // TODO(leofang): revisit this part when complex support is added to cupy.histogram()
@@ -981,7 +981,7 @@ struct _cub_histogram_range {
         // TODO(leofang): check if hipCUB has the same bug or not
 
         // if (n_samples < (1ULL << 31)) {
-            int num_samples = n_samples;
+            size_t num_samples = n_samples;
             DeviceHistogram::HistogramRange(workspace, workspace_size, static_cast<h_sampleT*>(input),
                 #ifndef CUPY_USE_HIP
                 static_cast<long long*>(output), n_bins, static_cast<h_binT*>(bins), num_samples, s);
@@ -1004,12 +1004,12 @@ struct _cub_histogram_range {
 struct _cub_histogram_even {
     template <typename sampleT>
     void operator()(void* workspace, size_t& workspace_size, void* input, void* output,
-        int& n_bins, int& lower, int& upper, size_t n_samples, cudaStream_t s) const
+        size_t& n_bins, size_t& lower, size_t& upper, size_t n_samples, cudaStream_t s) const
     {
         #ifndef CUPY_USE_HIP
         // Ugly hack to avoid specializing numerical types
-        typedef typename std::conditional<std::is_integral<sampleT>::value, sampleT, int>::type h_sampleT;
-        int num_samples = n_samples;
+        typedef typename std::conditional<std::is_integral<sampleT>::value, sampleT, size_t>::type h_sampleT;
+        size_t num_samples = n_samples;
         static_assert(sizeof(long long) == sizeof(intptr_t), "not supported");
         DeviceHistogram::HistogramEven(workspace, workspace_size, static_cast<h_sampleT*>(input),
             static_cast<long long*>(output), n_bins, lower, upper, num_samples, s);
@@ -1026,7 +1026,7 @@ struct _cub_histogram_even {
 /* -------- device reduce -------- */
 
 void cub_device_reduce(void* workspace, size_t& workspace_size, void* x, void* y,
-    int num_items, cudaStream_t stream, int op, int dtype_id)
+    size_t num_items, cudaStream_t stream, int op, int dtype_id)
 {
     switch(op) {
     case CUPY_CUB_SUM:      return dtype_dispatcher(dtype_id, _cub_reduce_sum(),
@@ -1045,7 +1045,7 @@ void cub_device_reduce(void* workspace, size_t& workspace_size, void* x, void* y
     }
 }
 
-size_t cub_device_reduce_get_workspace_size(void* x, void* y, int num_items,
+size_t cub_device_reduce_get_workspace_size(void* x, void* y, size_t num_items,
     cudaStream_t stream, int op, int dtype_id)
 {
     size_t workspace_size = 0;
@@ -1057,10 +1057,9 @@ size_t cub_device_reduce_get_workspace_size(void* x, void* y, int num_items,
 /* -------- device segmented reduce -------- */
 
 void cub_device_segmented_reduce(void* workspace, size_t& workspace_size,
-    void* x, void* y, int num_segments, int segment_size,
+    void* x, void* y, size_t num_segments, size_t segment_size,
     cudaStream_t stream, int op, int dtype_id)
 {
-    // CUB internally use int for offset...
     // This iterates over [0, segment_size, 2*segment_size, 3*segment_size, ...]
     #ifndef CUPY_USE_HIP
     thrust::counting_iterator<int> count_itr(0);
@@ -1089,7 +1088,7 @@ void cub_device_segmented_reduce(void* workspace, size_t& workspace_size,
 }
 
 size_t cub_device_segmented_reduce_get_workspace_size(void* x, void* y,
-    int num_segments, int segment_size,
+    size_t num_segments, size_t segment_size,
     cudaStream_t stream, int op, int dtype_id)
 {
     size_t workspace_size = 0;
@@ -1102,7 +1101,7 @@ size_t cub_device_segmented_reduce_get_workspace_size(void* x, void* y,
 /* -------- device scan -------- */
 
 void cub_device_scan(void* workspace, size_t& workspace_size, void* x, void* y,
-    int num_items, cudaStream_t stream, int op, int dtype_id)
+    size_t num_items, cudaStream_t stream, int op, int dtype_id)
 {
     switch(op) {
     case CUPY_CUB_CUMSUM:
@@ -1116,7 +1115,7 @@ void cub_device_scan(void* workspace, size_t& workspace_size, void* x, void* y,
     }
 }
 
-size_t cub_device_scan_get_workspace_size(void* x, void* y, int num_items,
+size_t cub_device_scan_get_workspace_size(void* x, void* y, size_t num_items,
     cudaStream_t stream, int op, int dtype_id)
 {
     size_t workspace_size = 0;
@@ -1128,7 +1127,7 @@ size_t cub_device_scan_get_workspace_size(void* x, void* y, int num_items,
 /* -------- device histogram -------- */
 
 void cub_device_histogram_range(void* workspace, size_t& workspace_size, void* x, void* y,
-    int n_bins, void* bins, size_t n_samples, cudaStream_t stream, int dtype_id)
+    size_t n_bins, void* bins, size_t n_samples, cudaStream_t stream, int dtype_id)
 {
     // TODO(leofang): support complex
     if (dtype_id == CUPY_TYPE_COMPLEX64 || dtype_id == CUPY_TYPE_COMPLEX128) {
@@ -1140,7 +1139,7 @@ void cub_device_histogram_range(void* workspace, size_t& workspace_size, void* x
                             workspace, workspace_size, x, y, n_bins, bins, n_samples, stream);
 }
 
-size_t cub_device_histogram_range_get_workspace_size(void* x, void* y, int n_bins,
+size_t cub_device_histogram_range_get_workspace_size(void* x, void* y, size_t n_bins,
     void* bins, size_t n_samples, cudaStream_t stream, int dtype_id)
 {
     size_t workspace_size = 0;
@@ -1150,7 +1149,7 @@ size_t cub_device_histogram_range_get_workspace_size(void* x, void* y, int n_bin
 }
 
 void cub_device_histogram_even(void* workspace, size_t& workspace_size, void* x, void* y,
-    int n_bins, int lower, int upper, size_t n_samples, cudaStream_t stream, int dtype_id)
+    size_t n_bins, size_t lower, size_t upper, size_t n_samples, cudaStream_t stream, int dtype_id)
 {
     #ifndef CUPY_USE_HIP
     return dtype_dispatcher(dtype_id, _cub_histogram_even(),
@@ -1158,8 +1157,8 @@ void cub_device_histogram_even(void* workspace, size_t& workspace_size, void* x,
     #endif
 }
 
-size_t cub_device_histogram_even_get_workspace_size(void* x, void* y, int n_bins,
-    int lower, int upper, size_t n_samples, cudaStream_t stream, int dtype_id)
+size_t cub_device_histogram_even_get_workspace_size(void* x, void* y, size_t n_bins,
+    size_t lower, size_t upper, size_t n_samples, cudaStream_t stream, int dtype_id)
 {
     size_t workspace_size = 0;
     cub_device_histogram_even(NULL, workspace_size, x, y, n_bins, lower, upper, n_samples,
